+++ R/utility_functions.R
@@ -688,62 +688,67 @@ StandardChromName <- function(df, file) {
   if (sum(grepl("GL", df[[1]])) > 0) {
     warning("In ", file, " ", sum(grepl("GL", df[[1]])), " row out of ",
             nrow(df), " had chromosome names that contain 'GL' and ", 
-            "were discarded. ")
+            "were removed. ",
+            "See the discarded variants in the return value for more details")
     df1 <- df[-grep("GL", df[[1]]), ]
     discarded.variants <- 
       dplyr::bind_rows(discarded.variants, df[grep("GL", df[[1]]), ])
   } else {
     df1 <- df
   }
-
+  
   # Is there any row in df whose Chromosome names have "KI"?
   if (sum(grepl("KI", df1[[1]])) > 0) {
     warning("In ", file, " ", sum(grepl("KI", df1[[1]])), " row out of ",
             nrow(df), " had chromosome names that contain 'KI' and ", 
-            "were discarded. ")
+            "were removed. ",
+            "See the discarded variants in the return value for more details")
     df2 <- df1[-grep("KI", df1[[1]]), ]
     discarded.variants <- 
       dplyr::bind_rows(discarded.variants, df1[grep("KI", df1[[1]]), ])
   } else {
     df2 <- df1
   }
-
+  
   # Is there any row in df whose Chromosome names have "random"?
   if (sum(grepl("random", df2[[1]])) > 0) {
     warning("In ", file, " ", sum(grepl("random", df2[[1]])), " row out of ",
             nrow(df), " had chromosome names that contain 'random' and ", 
-            "were discarded. ")
+            "were removed. ",
+            "See the discarded variants in the return value for more details")
     df3 <- df2[-grep("random", df2[[1]]), ]
     discarded.variants <- 
       dplyr::bind_rows(discarded.variants, df2[grep("random", df2[[1]]), ])
   } else {
     df3 <- df2
   }
-
+  
   # Is there any row in df whose Chromosome names are "Hs37D5"?
   if (sum(grepl("^Hs", df3[[1]])) > 0) {
     warning("In ", file, " ", sum(grepl("^Hs", df3[[1]])), " row out of ",
             nrow(df), " had chromosome names that contain 'Hs' and ", 
-            "were discarded. ")
+            "were removed. ",
+            "See the discarded variants in the return value for more details")
     df4 <- df3[-grep("^Hs", df3[[1]]), ]
     discarded.variants <- 
       dplyr::bind_rows(discarded.variants, df3[-grep("^Hs", df3[[1]]), ])
   } else {
     df4 <- df3
   }
-
+  
   # Is there any row in df whose Chromosome names contain "M"?
   if (sum(grepl("M", df4[[1]])) > 0) {
     warning("In ", file, " ", sum(grepl("M", df4[[1]])), " row out of ",
             nrow(df), " had chromosome names that contain 'M' and ", 
-            "were discarded. ")
+            "were removed. ",
+            "See the discarded variants in the return value for more details")
     df5 <- df4[-grep("M", df4[[1]]), ]
     discarded.variants <- 
       dplyr::bind_rows(discarded.variants, df4[grep("M", df4[[1]]), ])
   } else {
     df5 <- df4
   }
-
+  
   # Remove the "chr" character in the Chromosome's name
   df5[, 1] <- sub(pattern = "chr", replacement = "", df5[[1]])